// types.ts
export interface FormData {
  // Header Info (ID is generated by Backend)
  interviewDate: string;
  interviewerName: string;

  // 1. Identificación
  country: string;
  healthCenter: string;

  // 2. Demográficos
  fullName: string;
  dob: string;
  age: string;
  weight: string;
  height: string;

  // 3. Salud y Diagnóstico
  lastPeriodDate: string;
  comorbidities: string[]; // Multi-select
  comorbiditiesOther: string;
  currentMeds: string;
  confirmedDiagnosis: string; // 'Si', 'No'
  diagnosticMethods: string[]; // Multi-select
  diagnosticMethodsOther: string;
  diagnosisDate: string;
  clinicalStage: string;
  histologicalType: string;
  currentTreatment: string; // 'Si', 'No'
  treatmentTypes: string[]; // Multi-select
  treatmentOther: string;

  // 5. Muestra biológica
  // sampleCountry removed as requested (same as 1.1)
  sampleCollectionDateTime: string;
  sampleReceptionDateTime: string;
  collectionConditions: string;
  storageConditions: string;

  [key: string]: string | string[];
}

export enum FormStatus {
  IDLE = 'IDLE',
  PROCESSING = 'PROCESSING',
  SENDING = 'SENDING',
  SUCCESS = 'SUCCESS',
  ERROR = 'ERROR'
}